organize components into directories
