+++ frontend/interactEM/src/components/agents/chip.tsx
@@ -1,9 +1,9 @@
 import Chip from "@mui/material/Chip"
 import Tooltip from "@mui/material/Tooltip"
-import type { Agent } from "../types/agent"
-import { getAgentStatusColor } from "../utils/statusColor"
-import AgentTooltip from "./agenttooltip"
-import { StatusDot } from "./statusdot"
+import type { Agent } from "../../types/agent"
+import { getAgentStatusColor } from "../../utils/statusColor"
+import { StatusDot } from "../statusdot"
+import AgentTooltip from "./tooltip"
 
 interface AgentChipProps {
   agent: Agent
