+++ frontend/interactEM/src/components/agents/launchbutton.tsx
@@ -20,7 +20,7 @@ import {
   type AgentCreateEvent,
   agentsLaunchAgentMutation,
   zAgentCreateEvent,
-} from "../client"
+} from "../../client"
 
 // TODO: in the future, consider using: https://github.com/dohomi/react-hook-form-mui/
 export const LaunchAgentButton = () => {
