+++ frontend/interactEM/src/components/composerpipelineflow.tsx
@@ -36,10 +36,10 @@ import { layoutNodes } from "../layout"
 import { fromPipelineJSON, toJSON } from "../pipeline"
 import { usePipelineStore } from "../stores"
 import { NodeType, type OperatorNodeTypes } from "../types/nodes"
-import ImageNode from "./imagenode"
+import ImageNode from "./nodes/image"
+import OperatorNode from "./nodes/operator"
+import TableNode from "./nodes/table"
 import type { OperatorMenuItemDragData } from "./operatormenu"
-import OperatorNode from "./operatornode"
-import TableNode from "./tablenode"
 
 export const edgeOptions = {
   type: "smoothstep",
