+++ frontend/interactEM/src/components/nodes/parametersbutton.tsx
@@ -4,7 +4,7 @@ import SettingsIcon from "@mui/icons-material/Settings"
 import { Box, IconButton, Modal } from "@mui/material"
 import type React from "react"
 import { useState } from "react"
-import type { OperatorParameter } from "../client"
+import type { OperatorParameter } from "../../client"
 import ParameterUpdater from "./parameterupdater"
 
 interface ParametersButtonProps {
