+++ frontend/interactEM/src/pages/composerpage.tsx
@@ -2,10 +2,10 @@ import BuildIcon from "@mui/icons-material/Build"
 import VisibilityIcon from "@mui/icons-material/Visibility"
 import { Box, IconButton, Tooltip } from "@mui/material"
 import { useEffect } from "react"
-import AgentsAccordion from "../components/agentsaccordion"
+import AgentsAccordion from "../components/agents/accordion"
 import ComposerPipelineFlow from "../components/composerpipelineflow"
 import { OperatorMenu } from "../components/operatormenu"
-import { PipelineHud } from "../components/pipelinehud"
+import { PipelineHud } from "../components/pipelines/hud"
 import { useActivePipeline } from "../hooks/useActivePipeline"
 import { useAgents } from "../hooks/useAgents"
 import useOperators from "../hooks/useOperators"
